feat(events): deep links
* Added a "Copy link" button to events/comments in the UI, which copies
the current URL with a hash containing the event ID.

* When TimelineFeed is mounted, it sees if there's a relevant hash in
the URL, and requests the server to return the page that contains that
event (instead of just requesting the first page).

* Added an argument to the request event `search` route to return the
page containing a given event instead of a specific numbered page. The
actual page number chosen is returned by the server (see
inveniosoftware/invenio-records-resources#656).
This is calculated by counting the number of records older than the
specified one and dividing by the page size.