feat(events): quote comments in replies
* Added a button in the comment action dropdown to qute that comment's
contents in the new comment. When clicked, it appends the comment's
contents to the RichEditor (inside a blockquote).

* Added a tooltip that renders above any text highlighted within
comments and is dynamically positioned to be in the midpoint of the
selection. The tooltip contains a "Quote" button and a "Copy" button.
When "Quote" is clicked, it performs the same action as above, but just
for the selected portion of text.

* Formatting is only kept in the quote for the first method, due to the
complexities of propagating HTML formatting for partial selections.

* The blockquote currently has no real pointer to the quoted comment,
and is freely editable by the user.